requirements.txt
flask==3.0.2
RPi.GPIO==0.7.1
mfrc522==0.0.7
ffmpeg-python==0.2.0
python-dotenv==0.19.0
pysmb==1.2.9.1
spidev==3.6
smbus2==0.4.1
RPLCD==1.4.0